+++ server/test/integration/BraveAlerterConfiguratorTest/getAlertSessionTest.js
@@ -9,9 +9,8 @@ const BraveAlerterConfigurator = require('../../../BraveAlerterConfigurator')
 
 describe('BraveAlerterConfigurator.js integration tests: getAlertSession', () => {
   beforeEach(async () => {
-    this.sessionState = CHATBOT_STATE.WAITING_FOR_DETAILS
+    this.sessionState = CHATBOT_STATE.COMPLETED
     this.sessionIncidentType = '2'
-    this.sessionNotes = 'sessionNotes'
     this.message = 'message'
     this.installationResponderPhoneNumber = '+17775558888'
     this.installationIncidentCategories = ['Cat1', 'Cat2', 'Cat3']
@@ -30,7 +29,6 @@ describe('BraveAlerterConfigurator.js integration tests: getAlertSession', () =>
     this.sessionId = session.id
     session.state = this.sessionState
     session.incidentType = this.sessionIncidentType
-    session.notes = this.sessionNotes
     await db.saveSession(session)
   })
 
@@ -48,7 +46,7 @@ describe('BraveAlerterConfigurator.js integration tests: getAlertSession', () =>
         this.sessionId,
         this.sessionState,
         this.sessionIncidentType,
-        this.sessionNotes,
+        undefined,
         'There has been a request for help from 701 . Please respond "Ok" when you have followed up on the call.',
         this.installationResponderPhoneNumber,
         ['0', '1', '2'],