+++ server/test/testingHelpers.js
@@ -15,8 +15,6 @@ function createTestSessionState() {
     'fakeCreatedAt',
     'fakeUpdatedAt',
     '1',
-    'fakeNotes',
-    'fakeFallbackTwilioState',
     null,
     new Date('2000-06-06T00:53:53.000Z'),
   )