+++ server/test/unit/BraveAlerterConfiguratorTest/alertSessionChangedCallbackTest.js
@@ -93,20 +93,6 @@ describe('BraveAlerterConfigurator.js unit tests: alertSessionChangedCallback', 
     expect(db.saveSession).to.be.calledWith(expectedSession, sandbox.any)
   })
 
-  it('if given alertState WAITING_FOR_DETAILS should update only alertState', async () => {
-    sandbox.stub(db, 'getSessionWithSessionId').returns(createTestSessionState())
-
-    const sessionId = 'ca6e85b1-0a8c-4e1a-8d1e-7a35f838d7bc'
-    const braveAlerterConfigurator = new BraveAlerterConfigurator()
-    const braveAlerter = braveAlerterConfigurator.createBraveAlerter()
-    await braveAlerter.alertSessionChangedCallback(new AlertSession(sessionId, CHATBOT_STATE.WAITING_FOR_DETAILS))
-
-    const expectedSession = createTestSessionState()
-    expectedSession.state = CHATBOT_STATE.WAITING_FOR_DETAILS
-
-    expect(db.saveSession).to.be.calledWith(expectedSession, sandbox.any)
-  })
-
   it('if given alertState COMPLETED should update only alertState', async () => {
     sandbox.stub(db, 'getSessionWithSessionId').returns(createTestSessionState())
 
@@ -127,40 +113,11 @@ describe('BraveAlerterConfigurator.js unit tests: alertSessionChangedCallback', 
     const sessionId = 'ca6e85b1-0a8c-4e1a-8d1e-7a35f838d7bc'
     const braveAlerterConfigurator = new BraveAlerterConfigurator()
     const braveAlerter = braveAlerterConfigurator.createBraveAlerter()
-    await braveAlerter.alertSessionChangedCallback(new AlertSession(sessionId, CHATBOT_STATE.WAITING_FOR_DETAILS, '0'))
-
-    const expectedSession = createTestSessionState()
-    expectedSession.state = CHATBOT_STATE.WAITING_FOR_DETAILS
-    expectedSession.incidentType = 'Cat0'
-
-    expect(db.saveSession).to.be.calledWith(expectedSession, sandbox.any)
-  })
-
-  it('if given alertState and details should update alertState and details', async () => {
-    sandbox.stub(db, 'getSessionWithSessionId').returns(createTestSessionState())
-
-    const sessionId = 'ca6e85b1-0a8c-4e1a-8d1e-7a35f838d7bc'
-    const braveAlerterConfigurator = new BraveAlerterConfigurator()
-    const braveAlerter = braveAlerterConfigurator.createBraveAlerter()
-    await braveAlerter.alertSessionChangedCallback(new AlertSession(sessionId, CHATBOT_STATE.COMPLETED, undefined, 'fakeDetails'))
+    await braveAlerter.alertSessionChangedCallback(new AlertSession(sessionId, CHATBOT_STATE.COMPLETED, '0'))
 
     const expectedSession = createTestSessionState()
     expectedSession.state = CHATBOT_STATE.COMPLETED
-    expectedSession.notes = 'fakeDetails'
-
-    expect(db.saveSession).to.be.calledWith(expectedSession, sandbox.any)
-  })
-
-  it('if given only fallback return message should update only fallback return message', async () => {
-    sandbox.stub(db, 'getSessionWithSessionId').returns(createTestSessionState())
-
-    const sessionId = 'ca6e85b1-0a8c-4e1a-8d1e-7a35f838d7bc'
-    const braveAlerterConfigurator = new BraveAlerterConfigurator()
-    const braveAlerter = braveAlerterConfigurator.createBraveAlerter()
-    await braveAlerter.alertSessionChangedCallback(new AlertSession(sessionId, undefined, undefined, undefined, 'fakeFallbackReturnMessage'))
-
-    const expectedSession = createTestSessionState()
-    expectedSession.fallBackAlertTwilioStatus = 'fakeFallbackReturnMessage'
+    expectedSession.incidentType = 'Cat0'
 
     expect(db.saveSession).to.be.calledWith(expectedSession, sandbox.any)
   })